set dst/lib as linker rpath for rbczmq_ext
Setting the linker -rpath option to ext/rbczmq/dst/lib/ will allow
"rake" to successfully compile and run the tests.

Using a relative path means this will not work in the following
cases:

1. When the path can be resolved from the current working directory
2. After gem install; once the libraries have been installed to the gem
   path.

Would it be possible to compute an absolute path within the GEM PATH
where the dst/lib will end up?

refs #39