bump czmq/libzmq, statically link against them
Statically link against libzmq, czmq and libsodium, to avoid issues with
dynamic linking of libzmq (see e.g. #43).

Whilst on it (mainly because the existing versions of czmq/libzmq were
being a PITA to get to link statically properly), also bump the versions
of libzmq and czmq.

With czmq now being at API v3, also adjust the extension to be
compatible with it. It still uses a fair number of deprecated APIs, but
it absolutely works with API v3.

Fixes #43
Fixes #45 (hopefully)
Fixes #47